fix lints and make similar to ibc
