+++ x/wasm/keeper/handler_plugin_test.go
@@ -304,6 +304,7 @@ func TestIBCRawPacketHandler(t *testing.T) {
 		},
 	}
 	for name, spec := range specs {
+		spec := spec
 		t.Run(name, func(t *testing.T) {
 			capturedPacket = nil
 			// when
@@ -386,6 +387,7 @@ func TestBurnCoinMessageHandlerIntegration(t *testing.T) {
 	}
 	parentCtx := ctx
 	for name, spec := range specs {
+		spec := spec
 		t.Run(name, func(t *testing.T) {
 			ctx, _ = parentCtx.CacheContext()
 			k.wasmVM = &wasmtesting.MockWasmer{ExecuteFn: func(codeID wasmvm.Checksum, env wasmvmtypes.Env, info wasmvmtypes.MessageInfo, executeMsg []byte, store wasmvm.KVStore, goapi wasmvm.GoAPI, querier wasmvm.Querier, gasMeter wasmvm.GasMeter, gasLimit uint64, deserCost wasmvmtypes.UFraction) (*wasmvmtypes.Response, uint64, error) {
